+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 1.21.0 (September 09, 2020)
 
 CHANGED: